Revert "Implement API Consuming in /transactions/report"
This reverts commit 5a464199fea8c21211f71335a0710a4b3fd6c276.

+++ src/main/java/net/mert/reportingapi/model/TransactionReport.java
@@ -10,7 +10,7 @@ import javax.validation.constraints.Max;
 @AllArgsConstructor
 @NoArgsConstructor
 public class TransactionReport {
-    private Long count, total;
+    private int count, total;
 
     @Max(3)
     private String currency;
